# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CommentDetail" : {
        "properties" : {
          "authorDisplayName" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "postId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "renderedHtml" : {
            "type" : "string"
          }
        },
        "required" : [ "authorDisplayName", "createdAt", "id", "postId", "renderedHtml" ],
        "type" : "object"
      },
...
      "ContentOperationConfirmation" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "message" ],
        "type" : "object"
      },
...
      "EditCommentHttpRequest" : {
        "properties" : {
          "markdownSource" : {
            "type" : "string"
          }
        },
        "required" : [ "markdownSource" ],
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Bundled HTTP API for the server-streampack distribution of Streampack. Includes blog, authentication, and related site endpoints. Advisory version headers are supported: clients may send 'Accept-Version' and responses include 'Content-Version' and resolved 'Accept-Version'.",
    "title" : "server-streampack API",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/comments/{id}" : {
      "delete" : {
        "operationId" : "deleteComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "hard",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentOperationConfirmation"
                }
              }
            },
            "description" : "Comment deleted"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Admin access required"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Comment not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Delete a comment (soft by default, hard with ?hard=true)",
        "tags" : [ "Admin - Comment Management" ]
      }
    },
...
    "/comments/{id}" : {
      "put" : {
        "operationId" : "editComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditCommentHttpRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentDetail"
                }
              }
            },
            "description" : "Comment updated"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authenticated"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not authorized or edit window expired"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Comment not found"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "Edit a comment within the edit window",
        "tags" : [ "Comments" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local development server",
    "url" : "http://localhost:8080"
  } ]
}